[MOVE/#323] domain type 패키지를 만듭니다.
