+++ app/src/main/java/com/sopt/clody/domain/type/Notification.kt
@@ -1,4 +1,4 @@
-package com.sopt.clody.domain
+package com.sopt.clody.domain.type
 
 enum class Notification {
     DIARY, DRAFT, REPLY
