+++ app/src/main/java/com/sopt/clody/presentation/ui/home/navigation/HomeNavigation.kt
@@ -5,7 +5,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptionsBuilder
 import androidx.navigation.compose.composable
 import androidx.navigation.toRoute
-import com.sopt.clody.presentation.ui.type.ReplyStatus
+import com.sopt.clody.domain.type.ReplyStatus
 import com.sopt.clody.presentation.ui.home.screen.HomeRoute
 import com.sopt.clody.presentation.utils.navigation.Route
 import java.time.LocalDate
